service_python/requirements.txt
fastapi
uvicorn[standard]
httpx

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
